Fix bug where hashing function would be biased
The previous hasing function would never produce permutations 4 or 5. Switched with an implementation which generates a pseudorandom number from (0, 1) then maps it to [0...5]

I also manually picked an offset to plug into the pseudorandom function that would result in most permutations being picked mostly evenly for the duration of the internship.